Refactor Feature from trait to struct
After working with and thinking about the types of features
that should exist, it seems like a lot of 'features' I was thinking
of are actually just metadata that is attached to a value.

In this case, I think it makes more sense for no generics around
Feature here, and just allow for some metadata struct. This is
more restrictive for extension later, but focusing on that seems
like putting the cart before the horse.